requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorlog==6.9.0
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.5.0
hikari==2.1.0
hikari-lightbulb==3.0.0a15
idna==3.10
ipython==8.29.0
jedi==0.19.2
lxml==5.3.0
matplotlib-inline==0.1.7
multidict==6.1.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.0
